Fix SentenceEdit
