+++ lib/pages/sentence/edit.dart
@@ -1,4 +1,7 @@
+import 'package:booqs_mobile/components/sentence/edit_screen.dart';
+import 'package:booqs_mobile/data/provider/sentence.dart';
 import 'package:booqs_mobile/data/remote/sentences.dart';
+import 'package:booqs_mobile/i18n/translations.g.dart';
 import 'package:booqs_mobile/models/dictionary.dart';
 import 'package:booqs_mobile/models/sentence.dart';
 import 'package:booqs_mobile/pages/sentence/show.dart';
@@ -13,8 +16,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_easyloading/flutter_easyloading.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-class SentenceEditPage extends ConsumerStatefulWidget {
-  const SentenceEditPage({Key? key}) : super(key: key);
+class SentenceEditPage extends ConsumerWidget {
+  const SentenceEditPage({super.key});
 
   static Future push(BuildContext context, int sentenceId) async {
     return Navigator.of(context)
@@ -22,149 +25,34 @@ class SentenceEditPage extends ConsumerStatefulWidget {
   }
 
   @override
-  SentenceEditPageState createState() => SentenceEditPageState();
-}
-
-class SentenceEditPageState extends ConsumerState<SentenceEditPage> {
-  Sentence? _sentence;
-  bool _isLoading = true;
-  bool _isRequesting = false;
-  // validatorを利用するために必要なkey
-  final _formKey = GlobalKey<FormState>();
-  final _originalController = TextEditingController();
-  final _translationController = TextEditingController();
-  final _explanationController = TextEditingController();
-
-  @override
-  void initState() {
-    super.initState();
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      final arguments = ModalRoute.of(context)!.settings.arguments as Map;
-      final int sentenceId = arguments['sentenceId'];
-      _loadSentence(sentenceId);
-    });
-  }
-
-  Future _loadSentence(int sentenceId) async {
-    final Map? resMap = await RemoteSentences.edit(sentenceId);
-    _isLoading = false;
-    if (resMap == null) {
-      return setState(() => _isLoading);
-    }
-    _sentence = Sentence.fromJson(resMap['sentence']);
-    _originalController.text = _sentence?.original ?? '';
-    _translationController.text = _sentence?.translation ?? '';
-    _explanationController.text = _sentence?.explanation ?? '';
-    setState(() => _isLoading);
-  }
-
-  @override
-  // widgetの破棄時にコントローラも破棄する。Controllerを使うなら必ず必要。
-  // 参考： https://api.flutter.dev/flutter/widgets/TextEditingController-class.html
-  void dispose() {
-    _originalController.dispose();
-    _translationController.dispose();
-    _explanationController.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    // 更新処理
-    Future save(sentence) async {
-      // 各Fieldのvalidatorを呼び出す
-      if (!_formKey.currentState!.validate()) return;
-      setState(() {
-        _isRequesting = true;
-      });
-
-      Map<String, dynamic> params = {
-        'id': sentence.id,
-        'original': _originalController.text,
-        'translation': _translationController.text,
-        'explanation': _explanationController.text
-      };
-      // 画面全体にローディングを表示
-      EasyLoading.show(status: 'loading...');
-      final Map? resMap = await RemoteSentences.update(params);
-      EasyLoading.dismiss();
-      setState(() {
-        _isRequesting = false;
-      });
-      if (!mounted) return;
-
-      if (resMap == null) {
-        const snackBar = SnackBar(content: Text('例文を更新できませんでした。'));
-        ScaffoldMessenger.of(context).showSnackBar(snackBar);
-      } else {
-        final sentence = Sentence.fromJson(resMap['sentence']);
-        final snackBar = SnackBar(content: Text('${resMap['message']}'));
-        ScaffoldMessenger.of(context).showSnackBar(snackBar);
-        SentenceShowPage.pushReplacement(context, sentence.id);
-      }
-    }
-
-    // メイン
-    Widget body(Sentence? sentence) {
-      if (_isLoading) return const LoadingSpinner();
-      if (sentence == null) return const Text('Sentence does not exists.');
-      final Dictionary? dictionary = sentence.dictionary;
-      if (dictionary == null) return const Text('Dictionary does not exists.');
-
-      return Form(
-        key: _formKey,
-        child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: <Widget>[
-              DictionaryName(dictionary: dictionary),
-              const SizedBox(height: 32),
-              SentenceForm(
-                originalController: _originalController,
-                translationController: _translationController,
-                explanationController: _explanationController,
-                dictionary: sentence.dictionary!,
-                isNew: false,
-                keyword: '',
-              ),
-              const SizedBox(height: 40),
-              SizedBox(
-                height: 48,
-                child: ElevatedButton.icon(
-                  style: ElevatedButton.styleFrom(
-                    backgroundColor: Colors.green,
-                    minimumSize: const Size(double.infinity,
-                        40), // 親要素まで横幅を広げる。参照： https://stackoverflow.com/questions/50014342/how-to-make-button-width-match-parent
-                  ),
-                  onPressed: _isRequesting
-                      ? null
-                      : () async {
-                          save(_sentence);
-                        },
-                  icon: const Icon(Icons.update, color: Colors.white),
-                  label: const Text(
-                    '更新する',
-                    style: TextStyle(fontWeight: FontWeight.bold, fontSize: 16),
-                  ),
-                ),
-              ),
-              const SizedBox(height: 64),
-              SentenceFormDestroyButton(sentence: sentence),
-              const SizedBox(height: 160),
-            ]),
-      );
-    }
+  Widget build(BuildContext context, WidgetRef ref) {
+    final arguments = ModalRoute.of(context)!.settings.arguments as Map;
+    final int sentenceId = arguments['sentenceId'];
 
     return Scaffold(
       appBar: AppBar(
-        title: const Text('例文の編集'),
+        title: Text(t.sentences.edit),
       ),
-      body: SingleChildScrollView(
-        child: Container(
-          margin: EdgeInsets.symmetric(
-              vertical: 24,
-              horizontal: ResponsiveValues.horizontalMargin(context)),
-          child: body(_sentence),
-        ),
+      body: Container(
+        margin: EdgeInsets.symmetric(
+            vertical: 24,
+            horizontal: ResponsiveValues.horizontalMargin(context)),
+        child: ref.watch(asyncSentenceFamily(sentenceId)).when(
+            data: (Sentence? sentence) {
+              if (sentence == null) {
+                return const Text('Sentence does not exist.');
+              }
+              final Dictionary? dictionary = sentence.dictionary;
+              if (dictionary == null) {
+                return const Text('Dictionary does not exist.');
+              }
+              return SentenceEditScreen(
+                sentence: sentence,
+                dictionary: dictionary,
+              );
+            },
+            error: (err, str) => Text('Errors: $err'),
+            loading: () => const LoadingSpinner()),
       ),
       bottomNavigationBar: const BottomNavbar(),
     );
